LayoutTests/http/tests/media/hls/track-webvtt-multitracks-expected.txt


EXPECTED (video.textTracks.length == '2') OK
RUN(video.textTracks[1].mode = "showing")
EVENT(cuechange)
EXPECTED (video.textTracks[1].cues.length >= '1') OK
END OF TEST
